requirements.txt
ffmpeg-python
opencv-python
numpy
matplotlib
fastapi
python-multipart
uvicorn